requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-admin-argon-dashboard==1.0.16
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-fontawesome-5==1.0.18
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-material-admin==1.8.6
django-solo==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.21.5
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.1.0
packaging==23.1
Pillow==9.5.0
pkginfo==1.9.6
Pygments==2.15.1
PyJWT==2.7.0
pytz==2023.3
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.29.0
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.5
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
twine==4.0.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webencodings==0.5.1
zipp==3.15.0
